feat: falls back to TIME_DAY if the mon info is NULL or the toggle is off

+++ include/pokenav.h
@@ -18,7 +18,7 @@ struct PokenavMatchCallEntry
 {
     bool8 isSpecialTrainer;
     u8 mapSec;
-    u16 headerId;
+    u32 headerId;
 };
 
 struct PokenavListItem
